chore: refactorized editors to a separate customHook
